WindowsFile: Add experimental support for DeletePending
This is a port of a fix I made for Fluent Bit some time ago.

https://github.com/fluent/fluent-bit/blob/master/plugins/in_tail/win32/stat.c#L87

Basically, Windows has a special file state called "DeletePending",
where 1) the file is still accessible, 2) but we're supposed to
handle it as if it's already gone.

This patch makes Fluentd respect this state, and should fix the bug
"Windows k8s pods never shutdown due to Fluentd".

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>